pyproject.toml
[tool.poetry]
name = "openplugin"
version = "0.0.31"
description = ""
authors = ["shrikant <shrikant.pm14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain = "0.1.7"
requests = "2.31.0"
uvicorn = "^0.22.0"
fastapi = "^0.95.2"
load-dotenv = "^0.1.0"
openai = "1.12.0"
pydantic = "1.10.14"
typer = { extras = ["all"], version = "^0.9.0" }
websockets = "^11.0.3"
litellm = "1.23.14"
google-cloud-aiplatform = "1.35.0"
boto3 = "^1.29.3"
loguru = "^0.7.2"
pydub = "^0.25.1"


[tool.poetry.group.docs.dependencies]
nbsphinx = "^0.9.2"
sphinx-autobuild = "^2021.3.14"
sphinx = "6.2.1"
sphinx-book-theme = "^1.0.1"
toml = "^0.10.2"
sphinx-copybutton = "^0.5.2"
sphinx-tabs = "^3.4.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
mypy = "^1.5.1"
poethepoet = "^0.24.0"
black = "^23.9.1"
ruff = "^0.0.291"
pytest-cov = "^4.1.0"


[tool.poetry.group.types.dependencies]
types-toml = "^0.10.8.7"
types-requests = "^2.31.0.6"
types-pyyaml = "^6.0.12.12"


[tool.ruff]
select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "I",  # isort
]

fixable = ["ALL"]

# Same as Black.
line-length = 88

[tool.mypy]
disallow_untyped_defs = "False"
ignore_missing_imports = "True"


[tool.coverage.run]
omit = [
    "tests/*",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
openplugin = "openplugin.main:app"


[tool.poe.tasks]
build_docs = "sphinx-build docs/source _build"
black = "black"
ruff = "ruff"
pytest.cmd = "py.test --durations=5 -W error::RuntimeWarning --cov --cov-config=.coveragerc --cov-report xml --cov-report term-missing:skip-covered"
mypy = "mypy . --pretty --show-error-codes"
fix = { shell = "poe black . && poe ruff --fix ." }
test = { shell = "poe black . --check --diff && poe ruff . && poe pytest && poe mypy" }
serve.script = "start_api_server:main" 
